requirements.txt
simple_term_menu
pandas
numpy
scikit-learn
tensorflow
matplotlib
openai
crewai
crewai-tools
langchain
langchain-community
langchain-openai
langchain-core
pydantic
fastapi
uvicorn
python-multipart
pydantic
email-validator
python-jose
passlib
bcrypt
seaborn
google-auth
google_auth_oauthlib
google-api-python-client
boto3
python-dotenv
pytest
httpx
pydantic-settings
